consolidate population functions
